+++ system/lib/pthread/emscripten_futex_wait.c
@@ -9,8 +9,11 @@
 #include "pthread_impl.h"
 #include "threading_internal.h"
 
-#include <assert.h>
 #include <emscripten/threading.h>
+#include <emscripten/console.h>
+
+#include <assert.h>
+#include <math.h>
 #include <errno.h>
 #include <math.h>
 #include <stdatomic.h>
@@ -123,6 +126,12 @@ static int futex_wait_main_browser_thread(volatile void* addr,
   return 0;
 }
 
+static double dummy() {
+  return INFINITY;
+}
+
+weak_alias(dummy, _emscripten_next_timer);
+
 int emscripten_futex_wait(volatile void *addr, uint32_t val, double max_wait_ms) {
   if ((((intptr_t)addr)&3) != 0) {
     return -EINVAL;
@@ -146,92 +155,64 @@ int emscripten_futex_wait(volatile void *addr, uint32_t val, double max_wait_ms)
     return ret;
   }
 
+  // Pass 0 here, which means we don't have access to the current time in this
+  // function.  This tells _emscripten_yield to call emscripten_get_now if (and
+  // only if) it needs to know the time.
+  _emscripten_yield(0);
+
+  DBG("emscripten_futex_wait ms=%f", max_wait_ms);
+
+  bool is_runtime_thread = emscripten_is_main_runtime_thread();
+  if (is_runtime_thread) {
+    max_wait_ms = fmin(max_wait_ms, fmax(0, _emscripten_next_timer()));
+  }
+
   // -1 (or any negative number) means wait indefinitely.
   int64_t max_wait_ns = ATOMICS_WAIT_DURATION_INFINITE;
   if (max_wait_ms != INFINITY) {
     max_wait_ns = (int64_t)(max_wait_ms * 1e6);
   }
-
 #ifdef __EMSCRIPTEN_PTHREADS__
-  // When building with pthread support there are two conditions under which we
-  // need to limit the amount of time we spend in atomic.wait.
-  // 1. We are the main runtime thread.  In this case we need to be able to
-  //    process proxied events from workers.  Note that this is not always
-  //    the same as being the main browser thread.  For example, when running
-  //    under node or when launching an emscripten-built program in a Web
-  //    Worker.  In this case we limit our wait slices to 1ms intervals.
-  // 2. When the current thread has async cancellation enabled.  In this case
-  //    we limit the wait duration to 100ms intervals.
-  int64_t wakeup_interval = 0;
-  bool is_runtime_thread = emscripten_is_main_runtime_thread();
-  if (is_runtime_thread) {
-    // If the current thread is the main runtime thread then only wait in 1ms slices.
-    wakeup_interval = 1 * 1000000;
-  }
-  else if (cancelable) {
-    // If the current thread is async cancellable then only wait in 100ms slices.
-    wakeup_interval = 100 * 1000000;
-  }
-
-  // When wakeup_interval is set, we use end_time to track the absolute
-  // time when the wait should end.
-  double end_time = 0;
-  if (wakeup_interval) {
-    if (max_wait_ms == INFINITY) {
-      max_wait_ns = wakeup_interval;
-    } else {
-      end_time = emscripten_get_now() + max_wait_ms;
-      max_wait_ns = MIN(max_wait_ns, wakeup_interval);
-    }
-  }
-
-  do {
-#endif
-    // Pass 0 here, which means we don't have access to the current time in this
-    // function.  This tells _emscripten_yield to call emscripten_get_now if (and
-    // only if) it needs to know the time.
-    _emscripten_yield(0);
-
 #ifdef EMSCRIPTEN_DYNAMIC_LINKING
-    // After the main thread queues dlopen events, it checks if the target threads
-    // are sleeping.
-    // If `sleeping` is set then the main thread knows that event will be
-    // processed after the sleep (before any other user code).  In this case the
-    // main thread does not wait for any kind of response form the thread.
-    // If `sleeping` is not set then we know we should wait for the thread process
-    // the queue, either from the call here directly after setting `sleeping` to
-    // 1, or from another callsite (e.g. the one in `emscripten_yield`).
-    if (!is_runtime_thread) {
-      self->sleeping = 1;
-      _emscripten_process_dlopen_queue();
-    }
+  if (!is_runtime_thread) {
+    self->sleeping = 1;
+    _emscripten_process_dlopen_queue();
+  }
 #endif
+  uintptr_t expected_null = 0;
+  if (atomic_compare_exchange_strong(&self->wait_addr, &expected_null, (uintptr_t)addr)) {
+    DBG("emscripten_futex_wait atomic.wait ns=%lld", max_wait_ns);
     ret = __builtin_wasm_memory_atomic_wait32((int*)addr, val, max_wait_ns);
+  } else {
+    DBG("emscripten_futex_wait skipping atomic.wait due to NOTIFY_BIT");
+    // CAS failed, NOTIFY_BIT must have been set.  In this case we don't
+    // actually wait at all.  Instead we behave as if we spuriously woke up
+    // right away.
+    assert(expected_null & NOTIFY_BIT);
+    ret = ATOMICS_WAIT_OK;
+  }
+
+  // Clear the wait_addr
+  DBG("emscripten_futex_wait done notify=%d cancelable=%d cancel=%d", !!(self->wait_addr & NOTIFY_BIT), cancelable, self->cancel);
+  self->wait_addr = 0;
 #ifdef EMSCRIPTEN_DYNAMIC_LINKING
-    if (!is_runtime_thread) {
-      self->sleeping = 0;
-      _emscripten_process_dlopen_queue();
-    }
+  if (!is_runtime_thread) {
+    self->sleeping = 0;
+    _emscripten_process_dlopen_queue();
+  }
 #endif
-#ifdef __EMSCRIPTEN_PTHREADS__
-    if (cancelable && ret == ATOMICS_WAIT_TIMED_OUT && self->cancel) {
-      __pthread_testcancel();
-      // If __pthread_testcancel does return here it means that canceldisable
-      // must be set to PTHREAD_CANCEL_MASKED.  In this case we emulate the
-      // behaviour of the futex syscall and return ECANCELLED here.
-      // See pthread_cond_timedwait.c for the only use of this flag.
-      emscripten_conditional_set_current_thread_status(EM_THREAD_STATUS_WAITFUTEX, EM_THREAD_STATUS_RUNNING);
-      return -ECANCELED;
-    }
-    if (wakeup_interval && max_wait_ms != INFINITY) {
-      double remainder_ms = end_time - emscripten_get_now();
-      if (remainder_ms <= 0) {
-        break;
-      }
-      max_wait_ns = MIN((int64_t)(remainder_ms * 1e6), wakeup_interval);
-    }
-  } while (wakeup_interval && ret == ATOMICS_WAIT_TIMED_OUT);
-#endif
+  if (cancelable && self->cancel) {
+    __pthread_testcancel();
+    // If __pthread_testcancel does return here it means that canceldisable
+    // must be set to PTHREAD_CANCEL_MASKED.  In this case we emulate the
+    // behaviour of the futex syscall and return ECANCELLED here.
+    // See pthread_cond_timedwait.c for the only use of this flag.
+    emscripten_conditional_set_current_thread_status(EM_THREAD_STATUS_WAITFUTEX, EM_THREAD_STATUS_RUNNING);
+    return -ECANCELED;
+  }
+#else // __EMSCRIPTEN_PTHREADS__
+  ret = __builtin_wasm_memory_atomic_wait32((int*)addr, val, max_wait_ns);
+#endif // __EMSCRIPTEN_PTHREADS__
 
   emscripten_conditional_set_current_thread_status(EM_THREAD_STATUS_WAITFUTEX, EM_THREAD_STATUS_RUNNING);